requirements.txt
python-dateutil
requests
colorama